requirements.txt
psycopg2==2.7.6.1
requests
lxml
beautifulsoup4
python-dateutil